Restructure app to be more organized, modify padding on title/curator. Allow for marquee if song title is too long.
