+++ AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="io.soundtrack"
-    android:versionCode="1"
+    android:versionCode="2"
     android:versionName="0.1" >
 
     <uses-sdk
@@ -16,7 +16,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="io.soundtrack.MainActivity"
+            android:name="io.soundtrack.activities.MainActivity"
             android:label="@string/app_name" 
             android:configChanges="orientation"
             android:screenOrientation="portrait">
