+++ res/layout/activity_main.xml
@@ -19,13 +19,23 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
-        android:padding="@dimen/activity_content_margin"
+        android:paddingLeft="8dp"
+        android:paddingRight="8dp"
+        android:paddingTop="8dp"
          >
 
         <TextView
             android:id="@+id/songTitle"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            android:textStyle="bold"
+            android:maxLines="1"
+            android:ellipsize="marquee"
+            android:marqueeRepeatLimit="marquee_forever"
+            android:scrollHorizontally="true"
+            android:focusable="true" 
+    		android:focusableInTouchMode="true"
+    		android:freezesText="true" />
     </LinearLayout>
 
     <LinearLayout
@@ -33,7 +43,9 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
-        android:padding="@dimen/activity_content_margin" >
+        android:paddingLeft="8dp"
+        android:paddingRight="8dp"
+        android:paddingBottom="8dp" >
 
             <TextView
                 android:layout_width="wrap_content"
@@ -52,7 +64,8 @@
         android:layout_height="0dp"
         android:layout_weight="1"
         android:drawSelectorOnTop="false"
-        android:padding="@dimen/activity_content_margin" />
+        android:paddingLeft="8dp"
+        android:paddingRight="8dp" />
 
     <TextView
         android:layout_width="match_parent"
