+++ src/io/soundtrack/common/Configuration.java
@@ -1,6 +1,6 @@
 // Copyright 2012 Google Inc. All Rights Reserved.
 
-package io.soundtrack;
+package io.soundtrack.common;
 
 /**
  * Static container class for holding a reference to your YouTube Developer Key.
